### 2017-01-09 #154 @whedon
[PRE REVIEW]: bíogo: a simple high-performance bioinformatics toolkit for the Go language
**Submitting author:** @kortschak (<a href="http://orcid.org/0000-0001-8295-2301">Robert Daniel Kortschak</a>)
**Repository:** <a href="https://github.com/biogo/biogo" target ="_blank">https://github.com/biogo/biogo</a>
**Version:** v1.0.0

Thanks for submitting your paper to JOSS kortschak. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
