### 2017-01-20 #154 @arfon
> 👋 @pjotrp - would you be willing to edit this submission for JOSS?

Friendly reminder on this @pjotrp 😁
