### 2017-01-21 3e6f8ac
feat/gene: add tests and fix documentation
