### 2017-01-23 #154 @pjotrp
@brentp would you be willing to review this paper?
