### 2017-01-23 #154 @brentp
I am, but I am a (small) contributor to this repo. Does that preclude me?
