### 2017-01-24 9c637b4
feat/gene: add types for UTR5, CDS, UTR3 (#43)
* feat/gene: add types for UTR5, CDS, UTR3

* feat/gene: add field doc comments and mark deprecated methods

* feat/gene: add tests and fix documentation
