### 2017-01-24 #167 @whedon
[REVIEW]: bíogo: a simple high-performance bioinformatics toolkit for the Go language
**Submitting author:** kortschak (<a href="http://orcid.org/0000-0001-8295-2301">Robert Daniel Kortschak</a>)
**Repository:** <a href="https://github.com/biogo/biogo" target ="_blank">https://github.com/biogo/biogo</a>
**Version:** v1.0.0

- [x] **Authorship:** Has the submitting author (kortschak) made major contributions to the software?
