### 2017-01-24 fde8a27
travis: update test matrix
