+++ .travis.yml
@@ -1,9 +1,10 @@
 language: go
 
 go:
- - 1.4.2
- - 1.5.3
- - 1.6
+ - 1.5.4
+ - 1.6.4
+ - 1.7.4
+ - 1.8rc2
 
 # Get deps, build, test, and ensure the code is gofmt'ed.
 script:
