### 2017-01-24 479fddc
complexity: split pre and post go1.7 Z-complexity tests
The flate compression algorithm changed in 1.7, so the values differ.
