### 2017-01-25 4734ac4
io/seqio/fai: retain old error
