### 2017-02-01 #167 @brentp
This LGTM to me based on the criteria above and here: http://joss.theoj.org/about#reviewer_guidelines

I'm unclear what more I should do.
