### 2017-02-01 #167 @pjotrp
Thanks @brentp for your review!

@kortschak the paper.md is short and concise, but I would suggest to add some references to similar projects (biopython, biojava, bioruby, biorust) as a clarification of what the software aims to do and give it some historic reference (it also helps people to find you!). You can also add particular strengths of this project (at this point in time) so as to advertise the work to new comers and (I dare say) some performance metrics to underwrite your claims. Also add references to documentation so that people who want to dig deeper have a single click.

Note that these are merely suggestions to strengthen the scientific side of the paper and make it easy on the reader to understand what you are doing. We can accept the paper *as is* - so if you want to leave it like this that is fine too.
