### 2017-02-01 #167 @kortschak
@pjotrp I can refer explicitly to the original bioRxiv paper as that more fully details the reasoning behind the project and refers to other projects. The link to that is http://biorxiv.org/content/early/2014/05/12/005033 Do you think that would be satisfactory?
