### 2017-02-01 #167 @pjotrp
Good idea. After that ping @arfon that we complete acceptance.
