### 2017-02-01 4131037
paper: add reference to bioRxiv paper
Also fix line breaks.
