### 2017-02-02 #167 @kortschak
@arfon I would like to hold off until the release of the next version of the host compiler as there is an outstanding test suite failure with the current release candidate that should be fixed in that release. When Go1.8 is release I will update the test matrix and tag a minor version change, pinging @brentp and @pjotrp to ensure things have been done correctly. Is that OK with you?
