### 2017-02-04 801c72f
paper: properly mark up references
