+++ paper/paper.bib
@@ -1,0 +1,7 @@
+@Article{biogo,
+  author  = {Kortschak, R. D. and Adelson, D. L.},
+  title   = {bíogo: a simple high-performance bioinformatics toolkit for the Go language},
+  journal = {bioRxiv},
+  year    = {2014},
+  doi     = {10.1101/005033},
+}
