### 2017-02-16 64f4681
travis: update test matrix
