+++ version.go
@@ -4,4 +4,4 @@
 
 package biogo
 
-const Version = "v1.0.0"
+const Version = "v1.0.1"
