### 2017-02-16 #167 @kortschak
@brentp @pjotrp The required changes have now been made.

 http://doi.org/10.5281/zenodo.292939
