### 2017-02-18 #167 @arfon
@brentp & @pjotrp - many thanks for reviewing and editing this submission ✨

@kortschak - your paper is now accepted into JOSS. The DOI is http://dx.doi.org/10.21105/joss.00167 🚀 ⚡️ 💥
